feat: 멤버 엔티티 id값 수정

+++ src/main/java/be/stepnote/member/repository/MemberRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import javax.swing.text.html.Option;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<Member, Integer>, MemberRepositoryCustom {
+public interface MemberRepository extends JpaRepository<Member, Long>, MemberRepositoryCustom {
 
 
     Boolean existsByUsername(String username);
